aktualizacja widokow, zmiana wartości real (Faktury zewnetrzne), dodanie widoku do faktor

+++ Szwalnia/Faktury_Zewnetrzne.cs
@@ -18,9 +18,9 @@ namespace Szwalnia
         public string Nr_Faktury { get; set; }
         public Nullable<int> ID_Grupa { get; set; }
         public string Nazwa_Firmy { get; set; }
-        public float Netto { get; set; }
-        public float Brutto { get; set; }
-        public float Podatek { get; set; }
+        public int Netto { get; set; }
+        public int Brutto { get; set; }
+        public int Podatek { get; set; }
     
         public virtual Grupa Grupa { get; set; }
     }
